Adding Schema DSL

+++ net.eldm.parent/net.eldm/xtend-set/net/eldm/gen/CleanGen.java
@@ -12,12 +12,19 @@ import org.eclipse.emf.mwe.core.lib.AbstractWorkflowComponent2;
 import org.eclipse.emf.mwe.core.monitor.ProgressMonitor;
 
 public class CleanGen extends AbstractWorkflowComponent2 {
+    Boolean all = false;
+    
+    public Boolean getAll() { return all; }
+    public void setAll(Boolean all) { this.all = all; }
+    
     @Override
     protected void invokeInternal(WorkflowContext ctx, ProgressMonitor monitor, Issues issues) {
         try {
             final Path src = Paths.get("./src/net/eldm/");
-                deleteFile(src,"EldmDslStandaloneSetup.xtend");
+                deleteFile(src, "EldmDslStandaloneSetup.xtend");
                 deleteFile(src, "EldmDslRuntimeModule.xtend");
+                deleteFile(src, "SchDslStandaloneSetup.xtend");
+                deleteFile(src, "SchDslRuntimeModule.xtend");
                 
                 deleteDir(src, "./generator");
                 deleteDir(src, "./scoping");
@@ -30,15 +37,24 @@ public class CleanGen extends AbstractWorkflowComponent2 {
                 deleteDir(xtendgen, "./validation");
                 
             final Path srcgen = Paths.get("./src-gen/net/eldm/");
-                deleteFile(srcgen,"AbstractEldmDslRuntimeModule.java");
-                deleteFile(srcgen,"EldmDslStandaloneSetupGenerated.java");
-                deleteFile(srcgen,"EldmDsl.xtextbin");
+                deleteFile(srcgen, "AbstractEldmDslRuntimeModule.java");
+                deleteFile(srcgen, "EldmDslStandaloneSetupGenerated.java");
+                deleteFile(srcgen, "AbstractSchDslRuntimeModule.java");
+                deleteFile(srcgen, "SchDslStandaloneSetupGenerated.java");
                 
                 deleteDir(srcgen, "./parser");
                 deleteDir(srcgen, "./scoping");
                 deleteDir(srcgen, "./serializer");
                 deleteDir(srcgen, "./services");
                 deleteDir(srcgen, "./validation");
+                
+                if (all) {
+                    deleteFile(srcgen, "SchDsl.xtextbin");
+                    deleteFile(srcgen, "EldmDsl.xtextbin");
+                    
+                    deleteDir(srcgen, "./eldmDsl");
+                    deleteDir(srcgen, "./schDsl");
+                }
         } catch (Throwable ex) {
             ex.printStackTrace();
             throw new RuntimeException(ex.getMessage());
